增加基站的管理接口

+++ wintop-carauction-service/src/main/java/com/wintop/ms/carauction/mapper/read/TblAuctionBoardReadDao.java
@@ -20,6 +20,6 @@ public interface TblAuctionBoardReadDao {
     /**
      * 根据主键查询记录
      */
-    TblAuctionBoard selectByPrimaryKey(Integer id);
+    TblAuctionBoard selectByPrimaryKey(Long id);
 
 }
